collect SNI impl into src/tlsx_sni.c
