+++ cmake/functions.cmake
@@ -802,6 +802,7 @@ function(generate_lib_src_list LIB_SOURCES)
                    src/ssl.c 
                    src/tls.c
                    src/tlsx_base.c
+                   src/tlsx_sni.c
               )
 
               if(BUILD_TLS13)
