+++ src/include.am
@@ -585,7 +585,8 @@ src_libwolfssl_la_SOURCES += \
                src/keys.c \
                src/ssl.c \
                src/tls.c \
-               src/tlsx_base.c
+               src/tlsx_base.c \
+               src/tlsx_sni.c
 
 if BUILD_TLS13
 src_libwolfssl_la_SOURCES += src/tls13.c
