+++ wolfssl.vcxproj
@@ -284,6 +284,7 @@
     <ClCompile Include="src\ssl.c" />
     <ClCompile Include="src\tls.c" />
     <ClCompile Include="src\tlsx_base.c" />
+    <ClCompile Include="src\tlsx_sni.c" />
     <ClCompile Include="src\tls13.c" />
     <ClCompile Include="wolfcrypt\src\aes.c" />
     <ClCompile Include="wolfcrypt\src\arc4.c" />
